# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FitPass HOPn API Documentation",
        "description": "API Documentation for the FitPass HOPn Platform",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main API Server"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/partners": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Get a public list of approved partners",
                "operationId": "2a56c17491763bc4f6cda27614536748",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Lima"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "gym",
                                "spa",
                                "club"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of approved partners"
                    }
                }
            }
        },
        "/api/v1/partners/{id}": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Get details for a single partner",
                "operationId": "a586c51c0a3a96d6c17f2bd2ce777d3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partner details"
                    },
                    "404": {
                        "description": "Partner not found"
                    }
                }
            }
        },
        "/api/v1/partners/{id}/classes": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Get a list of all classes offered by a single partner",
                "operationId": "f78667bfa50bbdf99f74d55f79e1cbf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of classes"
                    },
                    "404": {
                        "description": "Partner not found"
                    }
                }
            }
        },
        "/api/v1/admin/partners": {
            "post": {
                "tags": [
                    "Super Admin - Partners"
                ],
                "summary": "Create a new partner",
                "operationId": "1ec9f97f47a526e3b88f24f3b9f04c2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "city",
                                    "address",
                                    "latitude",
                                    "longitude",
                                    "contact_email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "FitZone Gym"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "gym",
                                            "spa",
                                            "club"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Lima"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "contact_email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partner created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/partners/{id}": {
            "put": {
                "tags": [
                    "Super Admin - Partners"
                ],
                "summary": "Update an existing partner",
                "operationId": "b8b0b8a0168569aaf69717c4a146f483",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Partner updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Super Admin - Partners"
                ],
                "summary": "Delete a partner",
                "operationId": "d9492272cd14f36787ebd83646da0b8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Partner deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}